[ refactor ] Move `TraverseNeutral` (previously `NeutralMappable`) to new file
